got fetch functions out of components
