+++ src/components/Dashboard/FormDashboard/FormReference.jsx
@@ -14,87 +14,18 @@ import { switchForm } from "../../../utils/switchOptions";
 import roles from "../../../utils/roles";
 import http from "../../../services/http-common";
 import trnaslationKeys from "../../../utils/translationKeys.json";
+import {
+  getCountries,
+  getReferencesByName,
+  postContribution,
+  putContribution,
+} from "../../../services/getData";
 
 // Context
 import { DataContext, UserContext } from "../../../App";
+
 // components
 import HeaderDashboard from "../ContentDashboard/HeaderDashboard";
-
-// Get countries list from external API
-const getCountries = async () => {
-  return await http()
-    .get("https://restcountries.com/v3.1/all")
-    .then((response) => {
-      if (response.status === 200) {
-        return response.data;
-      }
-    })
-    .then((response) =>
-      response
-        .map((countryResult) => ({
-          value: countryResult.translations.fra.common.toLowerCase(),
-          name: countryResult.translations.fra.common.toLowerCase(),
-          label: countryResult.translations.fra.common,
-        }))
-        .sort((a, b) => a.value.localeCompare(b.value))
-    );
-};
-
-// Reuse of the search function of the SearchResult component to find similar references
-const getSearchReferences = async () => {
-  return await http()
-    .get(`search/reference-name`)
-    .then((result) => {
-      if (result.status === 200) {
-        return result.data;
-      }
-    })
-    .then((data) =>
-      data.search.map((reference) => ({
-        id: reference.id,
-        name: reference.reference_name,
-      }))
-    )
-    .catch((_) => {
-      return [];
-    });
-};
-
-// Requests to the API to send a contribution
-const postContribution = async (contribution, token) => {
-  return await http(token)
-    .post("references", contribution)
-    .then((response) => {
-      if (response.status === 202) {
-        return false;
-      }
-    })
-    .catch(({ response }) => {
-      return response.data.error;
-    });
-  // .catch((error) => error.response.data.error);
-};
-
-// Requests to the API to update a contribution
-const putContribution = async (contribution, token) => {
-  if (Object.keys(contribution).length > 0) {
-    return await http(token)
-      .put(`references/${contribution.reference_id}`, {
-        ...contribution,
-        reference_status: 1,
-      })
-      .then((response) => {
-        if (response.status === 202) {
-          return true;
-        }
-      })
-      .catch(() => {
-        return false;
-      });
-  }
-
-  return false;
-};
 
 //Displays the form for adding / modifying references
 export default function FormReference({ reference }) {
@@ -138,6 +69,7 @@ export default function FormReference({ reference }) {
       themesIds.push(option.id);
     });
   }, [selectedOptions, themesIds]);
+
   const onSubmit = async ({ reference_name, reference_date }) => {
     const contribution = {
       reference_id: reference ? reference.id : null,
@@ -173,10 +105,6 @@ export default function FormReference({ reference }) {
     (async () => setCountries(await getCountries()))();
   }, []);
 
-  // get the all Reference name from  getSearchReferences function and set into suggestName state
-  useEffect(() => {
-    (async () => setSuggestName(await getSearchReferences()))();
-  }, []);
   useEffect(() => {
     (async () => setCountries(await getCountries()))();
   }, []);
@@ -188,6 +116,9 @@ export default function FormReference({ reference }) {
   useEffect(() => {
     setMatchReferenceName("");
     setReferencesFound(referenceNameInput);
+
+    // get the all Reference name from  getReferencesByName function and set into suggestName state
+    (async () => setSuggestName(await getReferencesByName()))();
   }, [referenceNameInput]);
 
   useEffect(() => {
